+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved the language localization for Chinese (`zh`)
 
+### Fixed
+
+- Resolved a validation error caused by empty strings in the asset profile details dialog of the admin control panel
+
 ## 3.39.0 - 2026-08-01
 
 ### Changed