+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 3.40.0 - 2026-08-02
+## Unreleased
 
 ### Added
 
 - Added the database model and endpoints to manage the stock splits of an asset profile (experimental)
+
+## 3.40.0 - 2026-08-02
 
 ### Changed